+++ docs/Changelog.md
@@ -16,6 +16,7 @@
 - new `ACTIVATION_FINISH` state in `VerificationState`
   -  when this status is reached, the activation needs to be finalized by calling `VerificationService.finishActivation` method.
 - Replaced `createPowerAuthActivationData(...)` with `createActivationBuilder(...)`, used internally by `activate(...)` and for advanced workflows.
+- `DocumentType` changed from an enum to a typealias of `String` to better accommodate dynamic configuration of scanned documents.
 
 ## 1.3.0 (Oct, 2024)